requirements.txt
python-dotenv
requests
pyperclip
watchdog